# strparse: speed up the hex parser somewhat