Around 2.3x speed-up parsing many large hexadecimal numbers:

original speed:

  120000000 nums in 10.56473 secs, 88.0 ns/nums, 11358547.605 nums/sec

patched speed:

  120000000 nums in 4.58182 secs, 38.2 ns/nums, 26190465.797 nums/sec

Decimal and octal parsing speeds are marginally faster as well.

Still very readable, compact and easy to follow.

Tweaks

- combine the max and the overflow check, gains 3ns/num
- one less indirection in the pointer, gains 3ns/num
- using the table lookup for hex nums, gains 5ns/num
- unfold the num_digit() macro, gains 3s/num
- use the hexasciitable unconditionally, gains 2ns/num
- use post-increment pointer in the table lookup, gains 1ns/num
- improved valid_digit() using the table for the hex case, gains 26 ns/num

Behavior changes:

- no longer returns STRE_TOO_BIG - only STRE_OVERFLOW
- does not move the char ** on error, which is probably better

Updated test 1664 to verify.

[benchmarking program](https://gist.github.com/bagder/d548ebff65d31feb6cc552c0b33b8a8f)