@dfandrich 2025-02-17
I don't see any protection from an out of bounds array access in the line
  int n = hexasciitable[*p++ - '0'] & 0x0f;