@bagder 2025-02-17
> I don't see any protection from an out of bounds array access in the line int n = hexasciitable[*p++ - '0'] & 0x0f;

`*p` is verified to be a "valid digit" already before, so it shouldn't reach outside of the hexasciitable