@dfandrich 2025-02-17
That's what you get when you try to read the unified diff directly…