@testclutch 2025-02-18
Analysis of PR #16374 at [aa90fffa](https://github.com/curl/curl/commit/aa90fffae8fc0312faa85b3f557b5dc99bba0d49):

[Test 363 failed](https://github.com/curl/curl/actions/runs/13380402003/job/37367928052#check-step-3), which has NOT been flaky recently, so **there could be a real issue in this PR**.

[Test 1501 failed](https://github.com/curl/curl/actions/runs/13380402010/job/37367853046#check-step-14), which has NOT been flaky recently, so **there could be a real issue in this PR**.

###### Generated by [Testclutch](https://testclutch.curl.se/)