@bagder 2025-02-18
I also measured decimal and octal parsing speeds, both the original and in this version. They both get a tiny bit faster, but barely noticeable.