@bagder 2025-02-18
The [dist / verify-out-of-tree-autotools-debug](https://github.com/curl/curl/actions/runs/13385003510/job/37382966442?pr=16374) job fails test 363 consistently in this PR even after repeated reruns but I cannot see what makes it happen...